requirements.txt
Flask
jsonify
request
slackclient